Added service layer

+++ Associate/src/main/java/com/revature/application/repository/AssociateRepository.java
@@ -4,9 +4,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.revature.application.model.Associate;
 
-public interface AssociateRepository extends JpaRepository<Associate, Integer> {
-	public Associate findByAssociateId(int id);
+public interface AssociateRepository extends JpaRepository<Associate, Long> {
+	public Associate findByAssociateId(Long id);
 	public Associate findByEmail(String email);
-	public Associate findByUnitId(int id);
-	public void deleteByAssociateId(int id);
+	public Associate findByUnitId(Long id);
+	public void deleteByAssociateId(Long id);
 }
